can't format time_t with %d; cast this small interval time to int
git-svn-id: https://svn.apache.org/repos/asf/httpd/httpd/trunk@1089624 13f79535-47bb-0310-9956-ffa450edef68

+++ server/mpm/winnt/service.c
@@ -149,7 +149,8 @@ void hold_console_open_on_error(void)
                 return;
         }
         remains = ((start + 30) - time(NULL));
-        sprintf (count, "%d...", remains);
+        sprintf(count, "%d...",
+                (int)remains); /* 30 or less, so can't overflow int */
         if (!SetConsoleCursorPosition(hConErr, coninfo.dwCursorPosition))
             return;
         if (!WriteConsole(hConErr, count, (DWORD)strlen(count), &result, NULL)
